requirements.txt
pytest>=8.3.5
neo4j>=5.28.1
langchain>=0.3.20
python-dotenv>=1.0.1
panel>=1.4.5
hvplot>=0.11.2
holoviews>=1.20.1
pandas>=2.2.3
folium>=0.19.5
duckdb>=1.2.1
langchain-community>=0.3.19
pydantic>=2.10.6
langchain-text-splitters>=0.3.3
langchain-core>=0.3.45
langchain-openai>=0.3.8
langchain-ollama>=0.2.3
faiss-cpu>=1.10.0
langgraph~=0.3.11